+++ src/main/java/com/poortorich/chat/realtime/payload/response/UserEnterResponsePayload.java
@@ -1,4 +1,4 @@
-package com.poortorich.chat.realtime.payload;
+package com.poortorich.chat.realtime.payload.response;
 
 import com.poortorich.chat.entity.enums.ChatMessageType;
 import com.poortorich.chat.entity.enums.MessageType;
