+++ src/main/java/com/poortorich/chat/service/ChatMessageService.java
@@ -5,11 +5,11 @@ import com.poortorich.chat.entity.ChatParticipant;
 import com.poortorich.chat.entity.Chatroom;
 import com.poortorich.chat.realtime.builder.SystemMessageBuilder;
 import com.poortorich.chat.realtime.builder.UserChatMessageBuilder;
-import com.poortorich.chat.realtime.payload.ChatroomClosedResponsePayload;
-import com.poortorich.chat.realtime.payload.UserChatMessagePayload;
-import com.poortorich.chat.realtime.payload.UserEnterResponsePayload;
-import com.poortorich.chat.realtime.payload.UserLeaveResponsePayload;
 import com.poortorich.chat.realtime.payload.request.ChatMessageRequestPayload;
+import com.poortorich.chat.realtime.payload.response.ChatroomClosedResponsePayload;
+import com.poortorich.chat.realtime.payload.response.UserChatMessagePayload;
+import com.poortorich.chat.realtime.payload.response.UserEnterResponsePayload;
+import com.poortorich.chat.realtime.payload.response.UserLeaveResponsePayload;
 import com.poortorich.chat.repository.ChatMessageRepository;
 import com.poortorich.user.entity.User;
 import java.util.List;
